+++ datacatalog/cloud-client/package.json
@@ -25,8 +25,8 @@
     "test": {
       "build": {
         "requiredEnvVars": [
-          "API_KEY",
-          "GCLOUD_PROJECT"
+          "GCLOUD_PROJECT",
+          "GCLOUD_DATASET_ID"
         ]
       }
     }
